[PIL-1195] fix PR bot
