+++ conf/application.conf
@@ -68,10 +68,6 @@ microservice {
   }
 }
 
-auditing {
-  enabled = false
-}
-
 contact-frontend {
   host = "http://localhost:9250"
   serviceId = "pillar2-submission-frontend"
