+++ project/AppDependencies.scala
@@ -14,14 +14,14 @@ object AppDependencies {
     "uk.gov.hmrc"       %% "play-frontend-hmrc-play-30" % "9.0.0",
     "uk.gov.hmrc"       %% "bootstrap-frontend-play-30" % bootstrapVersion,
     "uk.gov.hmrc.mongo" %% "hmrc-mongo-play-30"         % hmrcMongoVersion,
-    "org.typelevel"     %% "cats-core"                  % "2.11.0"
+    "org.typelevel"     %% "cats-core"                  % "2.12.0"
   )
 
   val test = Seq(
     "uk.gov.hmrc"            %% "bootstrap-test-play-30"   % bootstrapVersion,
     "uk.gov.hmrc.mongo"      %% "hmrc-mongo-test-play-30"  % hmrcMongoVersion,
-    "org.scalatestplus.play" %% "scalatestplus-play"       % "5.1.0",
-    "io.github.wolfendale"   %% "scalacheck-gen-regexp"    % "0.1.3",
+    "org.scalatestplus.play" %% "scalatestplus-play"       % "7.0.1",
+    "io.github.wolfendale"   %% "scalacheck-gen-regexp"    % "1.1.0",
     "org.scalatestplus"      %% "scalatestplus-scalacheck" % "3.1.0.0-RC2",
     "uk.gov.hmrc"            %% "bootstrap-test-play-30"   % bootstrapVersion,
     "com.vladsch.flexmark"    % "flexmark-all"             % "0.62.2",
@@ -29,7 +29,7 @@ object AppDependencies {
     "org.seleniumhq.selenium" % "selenium-java"            % seleniumVersion,
     "org.seleniumhq.selenium" % "htmlunit-driver"          % "3.64.0",
     "com.danielasfregola"    %% "random-data-generator"    % "2.9",
-    "org.mockito"            %% "mockito-scala-scalatest"  % "1.17.29"
+    "org.mockito"            %% "mockito-scala-scalatest"  % "1.17.37"
   ).map(_ % Test)
 
   val it: Seq[ModuleID] = Seq(
