+++ project/build.properties
@@ -1,2 +1,2 @@
-sbt.version=1.9.7
+sbt.version=1.9.9
 hmrc-frontend-scaffold.version=0.34.0
